pyproject.toml
[tool.poetry]
name = "poster"
version = "0.1.0"
description = ""
authors = ["Roman Okatev <43115066+Nebulawalker@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.1"
environs = "^9.5.0"
pillow = "^10.0.0"
django-admin-sortable2 = "^2.1.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
